requirements.txt
absl-py==2.1.0
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
defusedxml==0.7.1
dm-tree==0.1.8
docopt==0.6.2
executing==2.0.1
fastjsonschema==2.19.1
Flask==3.0.2
flatbuffers==23.5.26
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.0
h5py==3.10.0
idna==3.6
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_pygments==0.3.0
keras==3.0.5
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
namex==0.0.7
nbclient==0.9.0
nbconvert==7.16.2
nbformat==5.9.2
numpy==1.26.4
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pandocfilters==1.5.1
parso==0.8.3
pickleshare==0.7.5
pillow==10.2.0
pipreqs==0.5.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.9.0.post0
pypiwin32
pyzmq==25.1.2
referencing==0.33.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
setuptools==69.1.1
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.0rc0
termcolor==2.4.0
tinycss2==1.2.1
tornado==6.4
traitlets==5.14.1
typing_extensions==4.10.0
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.1
wheel==0.42.0
wrapt==1.16.0
yarg==0.1.9